Adding ability to output JSON values to screen.
Will need to add implement stringer interface for Places type. Also need
to remove unecessary fields from current setup.
